requirements.txt
discord.py
python-dotenv
google-genai
chromadb
langgraph
langchain-google-genai